Packaging Refactor Part 3
Last part, and everything is working as expected. (No tests)

Added more documentation to the AuthenticationFilter and move the remaining
classes to their respective packages.

Refactor Summary:
- Moved all DTOs to dto package
- Moved all Models to model package
- Moved all Repositories to repository package
- Moved all Services to service package
- Moved all Controllers to controller package
- Moved all Configs to config package

Code Summary:
- Added a public endpoint for itineraries
- Successfully adding a new itinerary to the database with relations to
    the user
- Added more mappings to the ItineraryController
- Added more documentation
