+++ src/main/java/com/beastlymc/triptimize/dto/request/AuthenticationRequest.java
@@ -1,4 +1,4 @@
-package com.beastlymc.triptimize.auth;
+package com.beastlymc.triptimize.dto.request;
 
 import lombok.AllArgsConstructor;
 import lombok.Builder;
